Touchups, support for automatic compilation of C++ bots

+++ bots/build-all.sh
@@ -1,13 +1,33 @@
 #!/bin/bash
 
 cd `dirname $0`
-for i in `ls _BUILTIN/`
-do
-	name=${i%%.c}
-	rm -rf $name/
-	bash ../setup-bot.sh $name
-	cp _BUILTIN/$i $name/$i
-done
+
+function resetup()
+{
+	for i in `ls $1/`
+	do
+		if [[ $i == *.c ]]
+		then
+			name=${i%%.c}
+			rm -rf $name/
+			echo Deleted $name
+			bash ../setup-bot.sh $name
+			cp $1/$i $name/$i
+		fi
+
+		if [[ $i == *.cpp ]]
+		then
+			name=${i%%.cpp}
+			rm -rf $name/
+			echo Deleted $name
+			bash ../setup-bot-cpp.sh $name
+			cp $1/$i $name/$i
+		fi
+	done
+}
+
+resetup _BUILTIN/
+resetup _CUSTOM/
 
 for i in `ls`
 do
